Work with dates and also format and indent
If date expression is undated beginDate and endDate are not coded.  In
this case, beginDate and endDate for digital objects will be the begin
date and end date from the unitdate at the next highest level of the
hierarchy.  It is also okay in EAD not to encode dateExpression when
begindate and enddate are present.  Digital objects need a single value
for display, so in these cases a date expression is created using the
dates in begin and end date.